fix: prevent MongoDB tests from deleting all data
- Replace delete_many({}) with targeted deletions based on user_id
- Only delete assets referenced by test transactions
- Ensure test cleanup only affects test-created data
- Fixes potential data loss in non-test environments

Files modified:
- tests/test_database_integration.py
- tests/test_cache_integration.py
- tests/test_all_test_data_files.py

+++ tests/test_cache_integration.py
@@ -26,16 +26,32 @@ def test_db(test_user_id):
     # Clean up before test
     db.column_mapping_cache.delete_many({"user_id": test_user_id})
     db.wallets.delete_many({"user_id": test_user_id})
-    db.assets.delete_many({})
-    db.transactions.delete_many({})
+    # Only delete assets that are referenced by test transactions
+    test_wallet_ids = [w["_id"] for w in db.wallets.find({"user_id": test_user_id})]
+    if test_wallet_ids:
+        test_asset_ids = [
+            t["asset_id"]
+            for t in db.transactions.find({"wallet_id": {"$in": test_wallet_ids}})
+        ]
+        if test_asset_ids:
+            db.assets.delete_many({"_id": {"$in": test_asset_ids}})
+    db.transactions.delete_many({"user_id": test_user_id})
 
     yield db
 
     # Clean up after test
     db.column_mapping_cache.delete_many({"user_id": test_user_id})
     db.wallets.delete_many({"user_id": test_user_id})
-    db.assets.delete_many({})
-    db.transactions.delete_many({})
+    # Only delete assets that are referenced by test transactions
+    test_wallet_ids = [w["_id"] for w in db.wallets.find({"user_id": test_user_id})]
+    if test_wallet_ids:
+        test_asset_ids = [
+            t["asset_id"]
+            for t in db.transactions.find({"wallet_id": {"$in": test_wallet_ids}})
+        ]
+        if test_asset_ids:
+            db.assets.delete_many({"_id": {"$in": test_asset_ids}})
+    db.transactions.delete_many({"user_id": test_user_id})
 
 
 @pytest.fixture
